Create generic multiselect component
