+++ assets/src/styles/staff-base.css
@@ -1,5 +1,3 @@
-@import "./_tom-select.css";
-
 body {
   border-block-start: 1rem solid var(--color-bn-ribbon-600);
 }
