+++ templates/_components/multiselect/multiselect-option.html
@@ -1,0 +1,3 @@
+<option value="{{ value }}" {% if is_active %}selected{% endif %}>
+  {{ name }}
+</option>
