+++ templates/_components/multiselect/multiselect.html
@@ -1,0 +1,5 @@
+<div class="multiselect">
+  <select data-multiselect data-placeholder="Select an org" multiple name="{{ name }}">
+    {{ children }}
+  </select>
+</div>
